Login/logout + auth done

+++ frontend/src/components/ArticleDetails.js
@@ -7,12 +7,22 @@ import { Box, Link } from "@mui/material";
 import { useArticlesContext } from "../hooks/useArticlesContext";
 import Nyhed1 from "../images/nyhed/nyhed-1.png";
 
+import { useAuthContext } from "../hooks/useAuthContext.js";
+
+
 const ArticleDetails = ({ article }) => {
   const { dispatch } = useArticlesContext();
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if(!user){
+      return
+    }
     const response = await fetch("/api/articles/" + article._id, {
       method: "DELETE",
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     });
     const json = await response.json();
